Rename expendableCourse to expandableCourse
